# docs/swagger.yaml
definitions:
...
  request.UserCreateRequest:
    properties:
      email:
        type: string
      name:
        type: string
      password:
        type: string
      role:
        type: integer
      username:
        type: string
    required:
    - email
    - name
    - password
    - role
    - username
    type: object
  request.UserDeleteRequest:
    properties:
      id:
        type: string
    required:
    - id
    type: object
  request.UserLoginRequest:
    properties:
      password:
        type: string
      username:
        type: string
    required:
    - password
    - username
    type: object
  request.UserRegisterRequest:
    properties:
      email:
        type: string
      name:
        type: string
      password:
        type: string
      username:
        type: string
    required:
    - email
    - name
    - password
    - username
    type: object
  request.UserUpdateRequest:
    properties:
      email:
        type: string
      id:
        type: string
      password:
        minLength: 6
        type: string
      role:
        type: integer
      username:
        maxLength: 20
        minLength: 3
        type: string
    required:
    - id
    type: object
info:
  contact: {}
  description: 没有其他东西啦，仅仅是所有的后端接口，不要乱用哦
  title: PgsHub Backend API
  version: "1.0"
paths:
...
  /api/users/:
    delete:
      consumes:
      - application/json
      parameters:
      - description: PgsToken
        in: header
        name: PgsToken
        required: true
        type: string
      - description: UserDeleteRequest
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/request.UserDeleteRequest'
      produces:
      - application/json
      responses: {}
      summary: 用户删除（Role≤1 或 (Request)UserId=(PgsToken)UserId）
      tags:
      - 用户
    get:
      consumes:
      - application/json
      parameters:
      - in: query
        name: email
        type: string
      - in: query
        name: id
        type: string
      - in: query
        name: name
        type: string
      - in: query
        name: page
        type: integer
      - in: query
        name: role
        type: integer
      - in: query
        name: size
        type: integer
      - in: query
        name: username
        type: string
      produces:
      - application/json
      responses: {}
      summary: 用户查询
      tags:
      - 用户
    post:
      consumes:
      - application/json
      parameters:
      - description: PgsToken
        in: header
        name: PgsToken
        required: true
        type: string
      - description: UserCreateRequest
        in: body
        name: 创建请求
        required: true
        schema:
          $ref: '#/definitions/request.UserCreateRequest'
      produces:
      - application/json
      responses: {}
      summary: 用户创建（Role<=1）
      tags:
      - 用户
    put:
      consumes:
      - application/json
      description: 若 Role>1，则自动忽略 UserUpdateRequest 中的 Role 属性
      parameters:
      - description: PgsToken
        in: header
        name: PgsToken
        required: true
        type: string
      - description: UserUpdateRequest
        in: body
        name: 更新请求
        required: true
        schema:
          $ref: '#/definitions/request.UserUpdateRequest'
      produces:
      - application/json
      responses: {}
      summary: 用户更新（Role≤1 或 (Request)UserId=(PgsToken)UserId）
      tags:
      - 用户
  /api/users/login:
    post:
      consumes:
      - application/json
      parameters:
      - description: UserLoginRequest
        in: body
        name: 登录请求
        required: true
        schema:
          $ref: '#/definitions/request.UserLoginRequest'
      produces:
      - application/json
      responses: {}
      summary: 用户登录
      tags:
      - 用户
  /api/users/logout:
    post:
      consumes:
      - application/json
      parameters:
      - description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses: {}
      summary: 用户登出
      tags:
      - 用户
  /api/users/register:
    post:
      consumes:
      - application/json
      parameters:
      - description: UserRegisterRequest
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/request.UserRegisterRequest'
      produces:
      - application/json
      responses: {}
      summary: 用户注册
      tags:
      - 用户
  /api/users/token/{token}:
    get:
      parameters:
      - description: token
        in: path
        name: token
        required: true
        type: string
      produces:
      - application/json
      responses: {}
      summary: Token 鉴定
      tags:
      - 用户
swagger: "2.0"